refactor: split up Context.tsx
