+++ src/components/Card/AddButton.tsx
@@ -1,6 +1,6 @@
 import { useContext } from "react";
 import Spinner from "../../icons/Spinner";
-import { Context } from "../../data/Context";
+import { Context } from "../../data/store/";
 import styles from "./card.module.css";
 import { Item } from "../../data/interfaces";
